test: rename the CSS tests
so that I can run them all with `-n/Nokogiri::CSS/` while I'm
developing the SyntaxTree integration.
